refactor(article): 게시글 상세조회 로직 service로 이관
