+++ src/main/kotlin/com/devfloor/untitled/article/presentation/ArticleController.kt
@@ -24,7 +24,7 @@ class ArticleController(
 ) {
     @GetMapping(value = ["/articles/{id}"])
     @ResponseStatus(value = HttpStatus.OK)
-    fun showById(@PathVariable id: Long): ArticleResponse = articleFacade.showByArticleId(id)
+    fun showById(@PathVariable id: Long): ArticleResponse = articleService.showByArticleId(id)
 
     @PostMapping(value = ["/articles"])
     @ResponseStatus(value = HttpStatus.CREATED)
